# Speed up setting the backref match object